requirements.txt
annotated-types==0.7.0
aws-lambda-powertools==3.4.0
bandit==1.8.0
black==24.10.0
boto3==1.35.81
boto3-stubs==1.35.86
botocore==1.35.81
botocore-stubs==1.35.86
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.6.9
docutils==0.16
flake8==7.1.1
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy==1.13.0
mypy-extensions==1.0.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pbr==6.1.0
platformdirs==4.3.6
pluggy==1.5.0
pyasn1==0.6.1
pycodestyle==2.12.1
pydantic_core==2.27.1
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-testdox==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.7.2
s3transfer==0.10.4
six==1.17.0
stevedore==5.4.0
types-awscrt==0.23.6
types-requests==2.32.0.20241016
types-s3transfer==0.10.4
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
